# docs/swagger.yaml
definitions:
...
  deliverysvc.Job:
    properties:
      createTime:
        type: string
      creator:
        type: integer
      destinationCity:
        type: string
      destinationCompanyName:
        description: Origin Details
        type: string
      destinationEmailAddress:
        type: string
      destinationFirstLineAddress:
        type: string
      destinationFirstName:
        type: string
      destinationLatitude:
        type: number
      destinationLongitude:
        type: number
      destinationNotes:
        type: string
      destinationPhoneNumber:
        type: string
      destinationPostcode:
        type: string
      destinationSecondLineAddress:
        type: string
      destinationSecondName:
        type: string
      destinationThirdLineAddress:
        type: string
      destinationTown:
        type: string
      id:
        type: string
      lastTimeModified:
        type: string
      orderStatus:
        type: integer
      originCity:
        type: string
      originCompanyName:
        description: Destination Details
        type: string
      originEmailAddress:
        type: string
      originFirstLineAddress:
        type: string
      originFirstName:
        type: string
      originLatitude:
        type: number
      originLongitude:
        type: number
      originNotes:
        type: string
      originPhoneNumber:
        type: string
      originPostcode:
        type: string
      originSecondLineAddress:
        type: string
      originSecondName:
        type: string
      originThirdLineAddress:
        type: string
      originTown:
        type: string
      shortID:
        type: string
      status:
        type: integer
      trackingURL:
        type: string
      worker:
        type: integer
      workerNotes:
        description: courier notes
        type: string
    type: object
  management.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  payments.ClientSecret:
    properties:
      clientSecret:
        type: string
    type: object
  payments.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/payments.ErrorResponseMessage'
    type: object
  payments.ErrorResponseMessage:
    properties:
      message:
        type: string
    type: object
  payments.PublicKey:
    properties:
      publishableKey:
        type: string
    type: object
...
  usersvc.DriverState:
    properties:
      active:
        type: boolean
      courierID:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /management/driver-state:
    get:
      consumes:
      - application/json
      description: Retrieves all the current driver statuses
      parameters:
      - description: 'token: test'
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usersvc.DriverState'
        "400":
          description: driver state
          schema:
            $ref: '#/definitions/management.ErrorResponse'
        "500":
          description: driver state
          schema:
            $ref: '#/definitions/management.ErrorResponse'
      summary: Retrieves all the current driver statuses
  /management/jobs:
    get:
      consumes:
      - application/json
      description: Retrieves all the jobs
      parameters:
      - description: 'token: test'
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/deliverysvc.Job'
        "400":
          description: job
          schema:
            $ref: '#/definitions/management.ErrorResponse'
        "500":
          description: job
          schema:
            $ref: '#/definitions/management.ErrorResponse'
      summary: Retrieves all the jobs
  /payment/config:
    get:
      consumes:
      - application/json
      description: retrieves the publishable key for stripe
      parameters:
      - description: 'token: test'
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Public key
          schema:
            $ref: '#/definitions/payments.PublicKey'
        "400":
          description: Config
          schema:
            $ref: '#/definitions/payments.ErrorResponse'
        "500":
          description: Config
          schema:
            $ref: '#/definitions/payments.ErrorResponse'
      summary: Gets the publishable key for stripe
  /payment/create-intent:
    post:
      consumes:
      - application/json
      description: retrieves the publishable key for stripe
      parameters:
      - description: 'token: test'
        in: header
        name: Authorization
        required: true
        type: string
      - description: job-id
        in: query
        name: job-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Public key
          schema:
            $ref: '#/definitions/payments.ClientSecret'
        "400":
          description: Client Secret
          schema:
            $ref: '#/definitions/payments.ErrorResponse'
        "500":
          description: Client Secret
          schema:
            $ref: '#/definitions/payments.ErrorResponse'
      summary: Gets the publishable key for stripe
...
swagger: "2.0"